refactor : @Query 로직 삭제, QueryDSL로 리펙토링

+++ src/main/java/com/acc/local/repository/ports/RefreshTokenRepositoryPort.java
@@ -13,5 +13,5 @@ public interface RefreshTokenRepositoryPort {
 
     Optional<RefreshTokenEntity> findByRefreshTokenAndIsActiveTrue(String refreshToken);
 
-    int deactivateByTokenAtomically(String token, LocalDateTime now);
+    long deactivateByTokenAtomically(String token, LocalDateTime now);
 }
